fix: url and mysql injection protection

+++ admin/includes/view_all_posts.php
@@ -167,17 +167,25 @@ if (isset($_POST['checkBoxArray'])) {
 <?php
 
 if (isset($_GET['delete'])) {
-  $the_post_id = $_GET['delete'];
-  $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
-  $delete_query = mysqli_query($connection, $query);
-  header("Location: posts.php");
+  if (isset($_SESSION['user_role'])) {
+    if ($_SESSION['user_role'] == 'admin') {
+      $the_post_id = $_GET['delete'];
+      $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
+      $delete_query = mysqli_query($connection, $query);
+      header("Location: posts.php");
+    }
+  }
 }
 
 if (isset($_GET['reset'])) {
-  $the_post_id = $_GET['reset'];
-  $query = "UPDATE posts SET post_views_count = 0 WHERE post_id =" . mysqli_real_escape_string($connection, $_GET['reset']);
-  $reset_query = mysqli_query($connection, $query);
-  header("Location: posts.php");
+  if (isset($_SESSION['user_role'])) {
+    if ($_SESSION['user_role'] == 'admin') {
+      $the_post_id = $_GET['reset'];
+      $query = "UPDATE posts SET post_views_count = 0 WHERE post_id =" . mysqli_real_escape_string($connection, $_GET['reset']);
+      $reset_query = mysqli_query($connection, $query);
+      header("Location: posts.php");
+    }
+  }
 }
 
 ?>
